Update main.c
changing data type in decleration from float to int

+++ main.c
@@ -1,5 +1,5 @@
 #include <stdio.h>
-float subtract(float num1,float num2);
+int subtract(float num1,float num2);
 float add(float num1,float num2);
 float multiplication(float num1, float num2);
 float division(float num1, float num2);
